+++ src/providers/graphql/tests.rs
@@ -2,6 +2,7 @@ use super::*;
 mod bookmarks;
 mod likes;
 mod list_members;
+mod list_metadata;
 mod lists;
 mod relationships;
 struct Mock {